requirements.txt
Django==4.2.7
python-decouple==3.8
Pillow==10.4.0
djangorestframework==3.14.0
django-filter==23.5
drf-yasg==1.21.7
django-cors-headers==4.3.1
setuptools==69.0.3
